+++ Makefile.pre.in
@@ -1015,7 +1015,6 @@ PYTHON_HEADERS= \
 		$(srcdir)/Include/internal/ceval.h \
 		$(srcdir)/Include/internal/gil.h \
 		$(srcdir)/Include/internal/mem.h \
-		$(srcdir)/Include/internal/pymalloc.h \
 		$(srcdir)/Include/internal/pystate.h \
 		$(srcdir)/Include/internal/warnings.h \
 		$(DTRACE_HEADERS)
